feat: 클럽 목록 페이지네이션 prefetch 적용 (#647)
* [#646] feat: 클럽 목록 페이지네이션에 다음 페이지 prefetch 적용


* [#646] fix: 필터 변경 시 페이지네이션 prefetch가 갱신되지 않는 문제 수정

useEffect 의존성 배열에서 searchParams가 빠져 있어, 페이지 번호가
그대로인 상태로 필터만 바뀌면 next-page prefetch가 구 필터 URL로
고정된 채 갱신되지 않는 문제가 있었습니다. buildPageUrl을
useCallback으로 안정화하고 searchParams.toString()을 의존성에
포함시켜 해결했습니다.


---------
